+++ python/python/glide/constants.py
@@ -42,6 +42,8 @@ TFunctionListResponse = List[
         Union[bytes, List[Mapping[bytes, Union[bytes, Set[bytes]]]]],
     ]
 ]
+# Response for function stats command on a single node.
+# The response holds a map with 2 keys: Current running function / script and information about it, and the engines and the information about it.
 TFunctionStatsSingleNodeResponse = Mapping[
     bytes,
     Union[
@@ -52,6 +54,8 @@ TFunctionStatsSingleNodeResponse = Mapping[
         ],
     ],
 ]
+# Full response for function stats command across multiple nodes.
+# It maps node address to the per-node response.
 TFunctionStatsFullResponse = Mapping[
     bytes,
     TFunctionStatsSingleNodeResponse,