+++ python/python/tests/test_async_client.py
@@ -7346,11 +7346,6 @@ class TestCommands:
         with pytest.raises(RequestError):
             await glide_client.bitop(BitwiseOperation.AND, destination, [set_key])
 
-    @pytest.mark.parametrize("cluster_mode", [True])
-    @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
-    async def test_shoham(self, glide_client: GlideClusterClient):
-        print(await glide_client.function_stats(RandomNode()))
-
     @pytest.mark.parametrize("cluster_mode", [True, False])
     @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
     async def test_bitfield(self, glide_client: TGlideClient):
@@ -8254,6 +8249,58 @@ class TestCommands:
             check_function_stats_response(
                 cast(TFunctionStatsSingleNodeResponse, node_response), [], 0, 0
             )
+
+    """@pytest.mark.parametrize("cluster_mode", [False])
+    @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP3])
+    async def test_function_stats_running_script(
+        self, request, cluster_mode, protocol, glide_client: TGlideClient
+    ):
+        min_version = "7.0.0"
+        if await check_if_server_version_lt(glide_client, min_version):
+            return pytest.mark.skip(reason=f"Redis version required >= {min_version}")
+
+        lib_name = f"mylib1C{get_random_string(5)}"
+        func_name = f"myfunc1c{get_random_string(5)}"
+        code = create_lua_lib_with_long_running_function(lib_name, func_name, 20, True)
+
+        # load the library
+        assert await glide_client.function_load(code, replace=True) == lib_name.encode()
+
+        # create a second client to run fcall
+        test_client = await create_client(
+            request, cluster_mode=cluster_mode, protocol=protocol, timeout=30000
+        )
+
+        async def endless_fcall_route_call():
+            await test_client.fcall_ro(func_name, arguments=[])
+
+        async def wait_and_function_stats():
+            # it can take a few seconds for FCALL to register as running
+            await asyncio.sleep(5)
+            result = await test_client.function_stats()
+            running_scripts = False
+            print(result)
+            for _, res in result.items():
+                if res.get(b"running_scripts"):
+                    if running_scripts:
+                        raise Exception("Already running script on a different node")
+                    running_scripts = True
+                    assert (
+                        res.get(b"running_scripts").get(b"name") == func_name.encode()
+                    )
+                    assert res.get(b"running_scripts").get(b"command") == [
+                        b"FCALL_RO",
+                        func_name.encode(),
+                        b"0",
+                    ]
+                    assert res.get("running_scripts").get("duration_ms") > 0
+
+            assert running_scripts
+
+        await asyncio.gather(
+            endless_fcall_route_call(),
+            wait_and_function_stats(),
+        )"""
 
     @pytest.mark.parametrize("cluster_mode", [True])
     @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
@@ -10203,5 +10250,6 @@ class TestScripts:
             await glide_client.invoke_script(script, keys=[key], args=[arg])
             == key.encode()
         )
+        await glide_client.close()
         await glide_client.close()
         await glide_client.close()